+++ examples/garden/garden/app.py
@@ -12,10 +12,10 @@ import uuid
 import click
 
 from gazebo.asgi import ASGIApp, Receive, Scope, Send, trust_all
-from gazebo.context import use_request_id
-from gazebo.ext.cli import SettingsGroup, default_log_config
-from gazebo.ext.uvicorn import serve_command
+from gazebo.ext.cli import SettingsGroup
+from gazebo.ext.uvicorn import default_log_config, serve_command
 from gazebo.ext.fastapi import GazeboApp, Overrides, Providers
+from gazebo.requestid import use_request_id
 from gazebo.tags import Tag, tags_metadata
 
 from .api import root_router